better names

+++ delay/delay2.h
@@ -1,6 +1,6 @@
 #include "avr/io.h"
 
-void delayMs(uint16_t i);
-void delayUs(uint16_t i);
+void delayMs(uint8_t i);
+void delayUs(uint8_t i);
 void delayOneUs();
 void delayOneMs();
